DatabaseManager-main/requirements.txt
mysql-connector-python
ruff